py_flask_trade/requirements.txt
# Flask CMS 项目依赖
# 核心框架
Lin-CMS==0.4.10
Flask==2.1.2
Werkzeug==2.1.2

# 数据库相关
Flask-SQLAlchemy
Flask-Migrate
flask-sqlacodegen==1.1.8

# 中间件和扩展
flask-cors==3.0.10
flask-socketio==5.1.1
flask_redis==0.4.0

# 图像处理
pillow==9.3.0

# 异步和WebSocket
gevent-websocket==0.10.1
blinker==1.4

# 配置管理
python-dotenv==0.19.2

# Redis
redis==4.5.1

# 开发工具
black==22.3.0
isort==5.10.1
flake8==4.0.1
pytest==6.2.5
coverage==6.3.2
watchdog==2.1.6
pre-commit==2.17.0 